requirements.txt
langchain==0.2.11
langchain-core==0.2.43
langchain-groq==0.1.10
python-dotenv==1.0.1
tiktoken==0.5.2

